test: add tests for IconPackPresenter
