+++ screens/iconpack/src/test/kotlin/dev/mslalith/focuslauncher/screens/iconpack/IconPackPresenterTest.kt
@@ -3,15 +3,14 @@ package dev.mslalith.focuslauncher.screens.iconpack
 import android.content.ComponentName
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
-import app.cash.turbine.TurbineContext
 import com.google.common.truth.Truth.assertThat
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
 import dagger.hilt.android.testing.HiltTestApplication
 import dev.mslalith.focuslauncher.core.common.appcoroutinedispatcher.test.TestAppCoroutineDispatcher
-import dev.mslalith.focuslauncher.core.common.model.getOrNull
+import dev.mslalith.focuslauncher.core.common.model.LoadingState
 import dev.mslalith.focuslauncher.core.data.repository.AppDrawerRepo
-import dev.mslalith.focuslauncher.core.data.test.repository.settings.FakeGeneralSettingsRepo
+import dev.mslalith.focuslauncher.core.data.repository.settings.GeneralSettingsRepo
 import dev.mslalith.focuslauncher.core.domain.apps.GetAllAppsOnIconPackChangeUseCase
 import dev.mslalith.focuslauncher.core.domain.apps.GetIconPackIconicAppsUseCase
 import dev.mslalith.focuslauncher.core.domain.iconpack.FetchIconPacksUseCase
@@ -22,14 +21,12 @@ import dev.mslalith.focuslauncher.core.model.app.App
 import dev.mslalith.focuslauncher.core.model.app.AppWithComponent
 import dev.mslalith.focuslauncher.core.model.appdrawer.AppDrawerItem
 import dev.mslalith.focuslauncher.core.testing.AppRobolectricTestRunner
-import dev.mslalith.focuslauncher.core.testing.CoroutineTest
 import dev.mslalith.focuslauncher.core.testing.TestApps
+import dev.mslalith.focuslauncher.core.testing.circuit.PresenterTest
 import dev.mslalith.focuslauncher.core.testing.disableAsSystem
-import dev.mslalith.focuslauncher.core.testing.extensions.assertFor
 import dev.mslalith.focuslauncher.core.testing.launcherapps.TestIconPackManager
 import dev.mslalith.focuslauncher.core.testing.toPackageNamed
-import dev.mslalith.focuslauncher.screens.iconpack.model.IconPackState
-import kotlinx.coroutines.flow.StateFlow
+import kotlinx.collections.immutable.ImmutableList
 import kotlinx.coroutines.runBlocking
 import org.junit.Before
 import org.junit.FixMethodOrder
@@ -44,7 +41,7 @@ import javax.inject.Inject
 @RunWith(AppRobolectricTestRunner::class)
 @Config(application = HiltTestApplication::class)
 @FixMethodOrder(value = MethodSorters.NAME_ASCENDING)
-class IconPackViewModelTest : CoroutineTest() {
+class IconPackPresenterTest : PresenterTest<IconPackPresenter, IconPackState>() {
 
     @get:Rule(order = 0)
     val hiltRule = HiltAndroidRule(this)
@@ -56,89 +53,94 @@ class IconPackViewModelTest : CoroutineTest() {
     lateinit var getIconPackIconicAppsUseCase: GetIconPackIconicAppsUseCase
 
     @Inject
+    lateinit var generalSettingsRepo: GeneralSettingsRepo
+
+    @Inject
     lateinit var appDrawerRepo: AppDrawerRepo
-
-    private val generalSettingsRepo = FakeGeneralSettingsRepo()
-    private val appCoroutineDispatcher = TestAppCoroutineDispatcher()
-
-    private lateinit var viewModel: IconPackViewModel
 
     private val testIconProvider = TestIconProvider
     private val testIconPackManager = TestIconPackManager()
     private val fetchIconPacksUseCase = FetchIconPacksUseCase(iconPackManager = testIconPackManager)
     private val loadIconPackUseCase = LoadIconPackUseCase(iconPackManager = testIconPackManager)
 
+    private val appCoroutineDispatcher = TestAppCoroutineDispatcher()
     private val allApps by lazy { TestApps.all.toPackageNamed().disableAsSystem() }
 
     @Before
     fun setup() {
         hiltRule.inject()
-        viewModel = IconPackViewModel(
-            getAllAppsOnIconPackChangeUseCase = getAllAppsOnIconPackChangeUseCase,
-            getIconPackIconicAppsUseCase = getIconPackIconicAppsUseCase,
-            fetchIconPacksUseCase = fetchIconPacksUseCase,
-            loadIconPackUseCase = loadIconPackUseCase,
-            generalSettingsRepo = generalSettingsRepo,
-            appCoroutineDispatcher = appCoroutineDispatcher
-        )
         runBlocking {
             appDrawerRepo.addApps(apps = allApps)
         }
     }
 
+    override fun presenterUnderTest() = IconPackPresenter(
+        navigator = navigator,
+        getAllAppsOnIconPackChangeUseCase = getAllAppsOnIconPackChangeUseCase,
+        getIconPackIconicAppsUseCase = getIconPackIconicAppsUseCase,
+        fetchIconPacksUseCase = fetchIconPacksUseCase,
+        loadIconPackUseCase = loadIconPackUseCase,
+        generalSettingsRepo = generalSettingsRepo,
+        appCoroutineDispatcher = appCoroutineDispatcher
+    )
+
     @Test
-    fun `01 - when an icon pack is selected, state must be updated with it's new value`() = runCoroutineTest {
+    fun `01 - when an icon pack is selected, state must be updated with it's new value`() = runPresenterTest {
         val selectedIconPackType = IconPackType.Custom(packageName = "com.test")
 
-        viewModel.iconPackState.assertFor(expected = IconPackType.System) { it.iconPackType }
+        assertThat(awaitItem().iconPackType).isNull()
 
-        viewModel.updateSelectedIconPackApp(iconPackType = selectedIconPackType)
+        val state = awaitItem()
+        assertThat(state.iconPackType).isEqualTo(IconPackType.System)
 
-        viewModel.iconPackState.assertFor(expected = selectedIconPackType) { it.iconPackType }
+        state.eventSink(IconPackUiEvent.UpdateSelectedIconPackApp(iconPackType = selectedIconPackType))
+
+        assertThat(awaitItem().iconPackType).isEqualTo(selectedIconPackType)
     }
 
     @Test
-    fun `02 - when an icon pack is saved, state must be updated with it's new value`() = runCoroutineTest {
+    fun `02 - when an icon pack is saved, state must be updated with it's new value`() = runPresenterTest {
         val selectedIconPackType = IconPackType.Custom(packageName = "com.test")
 
-        viewModel.iconPackState.assertFor(expected = IconPackType.System) { it.iconPackType }
+        assertThat(awaitItem().iconPackType).isNull()
 
-        viewModel.updateSelectedIconPackApp(iconPackType = selectedIconPackType)
-        viewModel.saveIconPackType()
+        val state = awaitItem()
+        assertThat(state.iconPackType).isEqualTo(IconPackType.System)
 
-        viewModel.iconPackState.assertFor(expected = selectedIconPackType) { it.iconPackType }
+        state.eventSink(IconPackUiEvent.UpdateSelectedIconPackApp(iconPackType = selectedIconPackType))
+        state.eventSink(IconPackUiEvent.SaveIconPack)
+
+        assertThat(awaitItem().iconPackType).isEqualTo(selectedIconPackType)
     }
 
     @Test
-    fun `03 - when icon pack is changed, all apps with icons should have selected icon pack icons`() = runCoroutineTest {
+    fun `03 - when icon pack is changed, all apps with icons should have selected icon pack icons`() = runPresenterTest {
         val selectedIconPackType = IconPackType.Custom(packageName = "com.test")
-
         testIconProvider.setIconColor(color = Color.CYAN)
-        viewModel.iconPackState.assertAllAppsWith(apps = allApps, iconPackType = IconPackType.System)
+
+        assertThat(awaitItem().allApps).isEqualTo(LoadingState.Loading)
+        assertThat(awaitItem().allApps).isEqualTo(LoadingState.Loading)
+
+        val state = awaitItem()
+        state.allApps.assertWith(expected = allApps, iconPackType = IconPackType.System)
 
         testIconProvider.setIconColor(color = Color.BLUE)
-        viewModel.updateSelectedIconPackApp(iconPackType = selectedIconPackType)
+        state.eventSink(IconPackUiEvent.UpdateSelectedIconPackApp(iconPackType = selectedIconPackType))
 
-        viewModel.iconPackState.assertAllAppsWith(apps = allApps, iconPackType = selectedIconPackType)
+        assertThat(awaitItem().allApps).isEqualTo(LoadingState.Loading)
+        assertThat(awaitItem().allApps).isEqualTo(LoadingState.Loading)
+        awaitItem().allApps.assertWith(expected = allApps, iconPackType = selectedIconPackType)
     }
 
-    context (TurbineContext)
-    private suspend fun StateFlow<IconPackState>.assertAllAppsWith(
-        apps: List<App>,
-        iconPackType: IconPackType
-    ) = with(testIconProvider) { assertAllAppsWith(expected = apps.toAppDrawerItems(iconPackType = iconPackType)) }
-}
+    private fun LoadingState<ImmutableList<AppDrawerItem>>.assertWith(expected: List<App>, iconPackType: IconPackType) {
+        check(this is LoadingState.Loaded)
+        assertThat(value.size).isEqualTo(expected.size)
 
-context (TurbineContext)
-private suspend fun StateFlow<IconPackState>.assertAllAppsWith(
-    expected: List<AppDrawerItem>
-) {
-    assertFor(
-        expected = expected,
-        valueFor = { it.allApps.getOrNull() },
-        compare = { a, b -> a compareWith b },
-        assertion = { assertThat(it compareWith expected).isTrue() }
-    )
+        val expectedWithIcon = with(testIconProvider) { expected.toAppDrawerItems(iconPackType = iconPackType) }
+        value.zip(expectedWithIcon) { a, b ->
+            assertThat(a compareWith b).isTrue()
+        }
+    }
 }
 
 context (TestIconProvider)
@@ -155,18 +157,6 @@ private fun List<App>.toAppDrawerItems(iconPackType: IconPackType): List<AppDraw
         ),
         color = null
     )
-}
-
-private infix fun List<AppDrawerItem>?.compareWith(appDrawerItems: List<AppDrawerItem>?): Boolean {
-    if (this == null && appDrawerItems == null) return true
-    if (this == null || appDrawerItems == null) return false
-    if (size != appDrawerItems.size) return false
-
-    val thisSortedList = sortedBy { it.app.packageName }
-    val thatSortedList = appDrawerItems.sortedBy { it.app.packageName }
-    return thisSortedList.zip(other = thatSortedList).all { (thisItem, thatItem) ->
-        thisItem compareWith thatItem
-    }
 }
 
 private infix fun AppDrawerItem.compareWith(appDrawerItem: AppDrawerItem): Boolean {
